+++ frontend/src/routes/index.tsx
@@ -1,7 +1,7 @@
 import { Navigate, Route, Routes } from 'react-router-dom';
 import AppLayout from '../presentation/components/layout/AppLayout';
 import Overview from '../presentation/pages/Overview';
-import EquipamentosPage from '../modules/assets/presentation/EquipamentosPage';
+import EquipmentList from '../pages/Equipment/EquipmentList';
 import UsuariosPage from '../modules/users/presentation/UsuariosPage';
 import WorkOrders from '../presentation/pages/WorkOrders';
 import Plans from '../presentation/pages/Plans';
@@ -27,7 +27,7 @@ const Router = () => {
       >
         <Route index element={<DashboardPage />} />
         <Route path="overview" element={<Overview />} />
-        <Route path="equipamentos" element={<EquipamentosPage />} />
+        <Route path="equipamentos" element={<EquipmentList />} />
         <Route path="usuarios" element={<UsuariosPage />} />
         <Route path="work-orders" element={<WorkOrders />} />
         <Route path="plans" element={<Plans />} />